requirements.txt
Django==2.2.4
gunicorn==19.9.0
psycopg2-binary==2.8.3
whitenoise==4.1.3
django-environ==0.4.5
black==19.3b0
djangorestframework==3.10.2
